2Task. Change "UITextField" in UserSettingsViewController to "ClearableTextField"
First of all, make all textfields in our QuizApp to be ClearableTextField instead of UITextField. Secondly, clear icon is visible all the time now. It does not matter if we are editing, if we have text in textField or not. Please make it appear only if we are editing textfield & textfield contains text (not placeholder text but real text) TIP: UITextFieldDelegate